+++ ignite/templates/typed/list/list.go
@@ -11,7 +11,6 @@ import (
 
 	"github.com/ignite/cli/v29/ignite/pkg/errors"
 	"github.com/ignite/cli/v29/ignite/pkg/gomodulepath"
-	"github.com/ignite/cli/v29/ignite/pkg/placeholder"
 	"github.com/ignite/cli/v29/ignite/pkg/protoanalysis/protoutil"
 	"github.com/ignite/cli/v29/ignite/pkg/xast"
 	"github.com/ignite/cli/v29/ignite/templates/typed"
@@ -29,7 +28,7 @@ var (
 )
 
 // NewGenerator returns the generator to scaffold a new type in a module.
-func NewGenerator(replacer placeholder.Replacer, opts *typed.Options) (*genny.Generator, error) {
+func NewGenerator(opts *typed.Options) (*genny.Generator, error) {
 	subMessages, err := fs.Sub(fsMessages, "files/messages")
 	if err != nil {
 		return nil, errors.Errorf("fail to generate sub: %w", err)
@@ -47,7 +46,7 @@ func NewGenerator(replacer placeholder.Replacer, opts *typed.Options) (*genny.Ge
 	g.RunFn(protoQueryModify(opts))
 	g.RunFn(typesKeyModify(opts))
 	g.RunFn(keeperModify(opts))
-	g.RunFn(clientCliQueryModify(replacer, opts))
+	g.RunFn(clientCliQueryModify(opts))
 
 	// Genesis modifications
 	genesisModify(opts, g)
@@ -56,7 +55,7 @@ func NewGenerator(replacer placeholder.Replacer, opts *typed.Options) (*genny.Ge
 		// Modifications for new messages
 		g.RunFn(protoTxModify(opts))
 		g.RunFn(typesCodecModify(opts))
-		g.RunFn(clientCliTxModify(replacer, opts))
+		g.RunFn(clientCliTxModify(opts))
 
 		if !opts.NoSimulation {
 			g.RunFn(moduleSimulationModify(opts))
@@ -385,7 +384,7 @@ func typesCodecModify(opts *typed.Options) genny.RunFn {
 	}
 }
 
-func clientCliTxModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
+func clientCliTxModify(opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
 		path := filepath.Join("x", opts.ModuleName, "module/autocli.go")
 		f, err := r.Disk.Find(path)
@@ -393,43 +392,53 @@ func clientCliTxModify(replacer placeholder.Replacer, opts *typed.Options) genny
 			return err
 		}
 
-		template := `{
-			RpcMethod: "Create%[2]v",
-			Use: "create-%[3]v %[6]s",
-			Short: "Create %[4]v",
-			PositionalArgs: []*autocliv1.PositionalArgDescriptor{%[5]s},
-		},
-		{
-			RpcMethod: "Update%[2]v",
-			Use: "update-%[3]v [id] %[6]s",
-			Short: "Update %[4]v",
-			PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "id"}, %[5]s},
-		},
-		{
-			RpcMethod: "Delete%[2]v",
-			Use: "delete-%[3]v [id]",
-			Short: "Delete %[4]v",
-			PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "id"}},
-		},
-		%[1]v`
-
-		replacement := fmt.Sprintf(
-			template,
-			typed.PlaceholderAutoCLITx,
+		createOption := fmt.Sprintf(
+			`{
+				RpcMethod: "Create%[1]v",
+				Use: "create-%[2]v %[4]s",
+				Short: "Create %[3]v",
+				PositionalArgs: []*autocliv1.PositionalArgDescriptor{%[5]s},
+			}`,
 			opts.TypeName.PascalCase,
 			opts.TypeName.Kebab,
 			opts.TypeName.Original,
-			opts.Fields.ProtoFieldNameAutoCLI(),
 			opts.Fields.CLIUsage(),
+			opts.Fields.ProtoFieldNameAutoCLI(),
 		)
-
-		content := replacer.Replace(f.String(), typed.PlaceholderAutoCLITx, replacement)
+		updateOption := fmt.Sprintf(
+			`{
+				RpcMethod: "Update%[1]v",
+				Use: "update-%[2]v [id] %[4]s",
+				Short: "Update %[3]v",
+				PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "id"}, %[5]s},
+			}`,
+			opts.TypeName.PascalCase,
+			opts.TypeName.Kebab,
+			opts.TypeName.Original,
+			opts.Fields.CLIUsage(),
+			opts.Fields.ProtoFieldNameAutoCLI(),
+		)
+		deleteOption := fmt.Sprintf(
+			`{
+				RpcMethod: "Delete%[1]v",
+				Use: "delete-%[2]v [id]",
+				Short: "Delete %[3]v",
+				PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "id"}},
+			}`,
+			opts.TypeName.PascalCase,
+			opts.TypeName.Kebab,
+			opts.TypeName.Original,
+		)
+		content, err := typed.AppendAutoCLITxOptions(f.String(), createOption, updateOption, deleteOption)
+		if err != nil {
+			return err
+		}
 		newFile := genny.NewFileS(path, content)
 		return r.File(newFile)
 	}
 }
 
-func clientCliQueryModify(replacer placeholder.Replacer, opts *typed.Options) genny.RunFn {
+func clientCliQueryModify(opts *typed.Options) genny.RunFn {
 	return func(r *genny.Runner) error {
 		path := filepath.Join("x", opts.ModuleName, "module/autocli.go")
 		f, err := r.Disk.Find(path)
@@ -437,27 +446,32 @@ func clientCliQueryModify(replacer placeholder.Replacer, opts *typed.Options) ge
 			return err
 		}
 
-		template := `{
-			RpcMethod: "List%[2]v",
-			Use: "list-%[3]v",
-			Short: "List all %[4]v",
-		},
-		{
-			RpcMethod: "Get%[2]v",
-			Use: "get-%[3]v [id]",
-			Short: "Gets a %[4]v by id",
-			Alias: []string{"show-%[3]v"},
-			PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "id"}},
-		},
-		%[1]v`
-		replacement := fmt.Sprintf(
-			template,
-			typed.PlaceholderAutoCLIQuery,
+		listOption := fmt.Sprintf(
+			`{
+				RpcMethod: "List%[1]v",
+				Use: "list-%[2]v",
+				Short: "List all %[3]v",
+			}`,
 			opts.TypeName.PascalCase,
 			opts.TypeName.Kebab,
 			opts.TypeName.Original,
 		)
-		content := replacer.Replace(f.String(), typed.PlaceholderAutoCLIQuery, replacement)
+		getOption := fmt.Sprintf(
+			`{
+				RpcMethod: "Get%[1]v",
+				Use: "get-%[2]v [id]",
+				Short: "Gets a %[3]v by id",
+				Alias: []string{"show-%[2]v"},
+				PositionalArgs: []*autocliv1.PositionalArgDescriptor{{ProtoField: "id"}},
+			}`,
+			opts.TypeName.PascalCase,
+			opts.TypeName.Kebab,
+			opts.TypeName.Original,
+		)
+		content, err := typed.AppendAutoCLIQueryOptions(f.String(), listOption, getOption)
+		if err != nil {
+			return err
+		}
 		newFile := genny.NewFileS(path, content)
 		return r.File(newFile)
 	}