+++ ignite/templates/typed/placeholders.go
@@ -1,8 +1,0 @@
-package typed
-
-//nolint:godot
-const (
-	// Placeholders AutoCLI
-	PlaceholderAutoCLIQuery = "// this line is used by ignite scaffolding # autocli/query"
-	PlaceholderAutoCLITx    = "// this line is used by ignite scaffolding # autocli/tx"
-)